add: 유저 앱과 가계부 앱의 모든 기능에 주석설명 추가 및 가계부 앱의 디렉토리 변경 #16
