+++ core/utils/get_obj.py
@@ -4,7 +4,20 @@ from account_books.models import AccountBook, AccountBookCategory, AccountBookLo
 from users.models         import User
 
 class GetAccountBook:
+    """
+    Assignee: 김동규
+    
+    param: account_book_id, user
+    return: obj, err
+    detail:
+      - 가계부 id를 통해 가계부 객체(정보)의 존재여부 확인
+      - 가계부 객체의 유저정보와 API를 호출한 유저의 정보를 대조
+    """
+    
     def get_book_n_check_error(account_book_id: int, user: User) -> Tuple[Any, str]:
+        """
+        가계부 객체/에러 확인
+        """
         try:
             book = AccountBook.objects\
                               .select_related('users')\
@@ -19,7 +32,20 @@ class GetAccountBook:
     
 
 class GetAccountBookCategory:
+    """
+    Assignee: 김동규
+    
+    param: account_book_category_id, user
+    return: obj, err
+    detail:
+      - 가계부 카테고리 id를 통해 가계부 카테고리 객체(정보)의 존재여부 확인
+      - 가계부 카테고리 객체의 유저정보와 API를 호출한 유저의 정보를 대조
+    """
+    
     def get_category_n_check_error(account_book_category_id: int, user: User) -> Tuple[Any, str]:
+        """
+        가계부 카테고리 객체/에러 확인
+        """
         try:
             category = AccountBookCategory.objects\
                                           .select_related('users')\
@@ -34,7 +60,20 @@ class GetAccountBookCategory:
     
 
 class GetAccountBookLog:
+    """
+    Assignee: 김동규
+    
+    param: account_book_log_id, user
+    return: obj, err
+    detail:
+      - 가계부 기록 id를 통해 가계부 기록 객체(정보)의 존재여부 확인
+      - 가계부 기록 객체의 유저정보와 API를 호출한 유저의 정보를 대조
+    """
+    
     def get_log_n_check_error(account_book_log_id: int, user: User) -> Tuple[Any, str]:
+        """
+        가계부 기록 객체/에러 확인
+        """
         try:
             log = AccountBookLog.objects\
                                 .select_related('books')\
